(Final):Finished proper tokenizer, flesh out final symbol table
