+++ src/compiler.rs
@@ -1,3 +1,1 @@
 pub mod lexanalysis;
-pub mod symboltable;
-pub mod fsa;
