+++ test.txt
@@ -1,0 +1,34 @@
+Token: CLASS      Symbol: Var
+Token: Pgm1       Symbol: NumLit
+Token: {          Symbol: OpenBracket
+Token: CONST      Symbol: Var
+Token: M          Symbol: Var
+Token: =          Symbol: Assign
+Token: 13         Symbol: NumLit
+Token: ,          Symbol: Comma
+Token: N          Symbol: Var
+Token: =          Symbol: Assign
+Token: 56         Symbol: NumLit
+Token: ;          Symbol: Semi
+Token: VAR        Symbol: Var
+Token: X          Symbol: Var
+Token: ,          Symbol: Comma
+Token: Y          Symbol: Var
+Token: ,          Symbol: Comma
+Token: Z          Symbol: Var
+Token: ;          Symbol: Semi
+Token: Y          Symbol: Var
+Token: =          Symbol: Assign
+Token: 97         Symbol: NumLit
+Token: ;          Symbol: Semi
+Token: X          Symbol: Var
+Token: =          Symbol: Assign
+Token: M          Symbol: Var
+Token: *          Symbol: Mop
+Token: N          Symbol: Var
+Token: +          Symbol: Addop
+Token: 18         Symbol: NumLit
+Token: -          Symbol: Addop
+Token: Y          Symbol: Var
+Token: ;          Symbol: Semi
+Token: }          Symbol: CloseBracket
